services/order_worker/requirements.txt
mysql-connector-python==8.3.0
psycopg2-binary==2.9.9
pika==1.3.2
